+++ web3/_utils/module_testing/eth_module.py
@@ -838,3 +838,10 @@ class EthModuleTest:
         node_id = '59daa26581d0acd1fce254fb7e85952f4c09d0915afd33d3886cd914bc7d283c'
         result = web3.eth.submitHashrate(5000, node_id)
         assert result is True
+
+    def test_eth_submitWork(self, web3):
+        nonce = 1
+        pow_hash = '0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef'
+        mix_digest = '0xD1FE5700000000000000000000000000D1FE5700000000000000000000000000'
+        result = web3.eth.submitWork(nonce, pow_hash, mix_digest)
+        assert result is False