+++ docs/releases.rst
@@ -38,6 +38,8 @@ Unreleased (latest source)
 --------------------------
   - Remove ``web3/utils`` directory in favor of ``web3/_utils``
     - `#1282 <https://github.com/ethereum/web3.py/pull/1282>`_
+  - Implement ``eth_submitHashrate`` and ``eth_submitWork`` JSONRPC endpoints.
+    - `#1280 <https://github.com/ethereum/web3.py/pull/1280>`_
 
 v5.0.0-alpha.7
 --------------